requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth>=2.23.0

